Use new JS Solr Utils library
The code has been updated to use the new JS Solr Utils library that
reduces the code and makes the partial be cleaner.

This commit contains the following changes:

Refactors code for fancyTree and collapsible plugin
The name of the plugins was changed on kmaps_engine and now the change
to those names has been reflected in this commit.
fancyTree was renamed to kmapsRelationsTree
collapsibleList was renamed to kmapsCollapsibleList

Update code to on related for summary
Cleaned code for new update on solr-utils library