Update code to take advantage of popup language support
Kmaps-popup now takes into account language, so this update takes
advantage of that.

Refactor code to use new implementation of relations tree.

A new version of the kmapsRelationsTree plugin has been published and we
refactored the code to take advantage of this change.

Fix localization for bo, dz, zh and updated the version in preparation for merge into master